backend/requirements-prod.txt
# Production requirements — used by Dockerfile.
# requirements.txt retains the full set (including dev tools) for local use.
# Excluded here (dev/test only):
#   pytest, pytest-asyncio, ruff, libcst
# Excluded here (optional/conditional import — gracefully disabled if absent):
#   onnxruntime  (both uses are try/except; CLIP dedup feature degrades, not crashes)

annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.1
feedparser==6.0.12
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.2.6
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
duckdb==1.5.0
playwright==1.58.0
pydantic==2.12.5
email-validator>=2.2.0
pydantic_core==2.41.5
pyee==13.0.1
Pygments==2.19.2
pypdf==6.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.8.3
starlette==0.52.1
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0

# Production stack
sqlmodel==0.0.34
sqlalchemy==2.0.38
asyncpg==0.30.0
aiosqlite>=0.19.0
redis==5.2.1
psycopg2-binary==2.9.10
scalar-fastapi>=1.8.0

# Intelligence Hub
curl_cffi>=0.7.0
h3>=3.7.7
aiofiles>=23.0
lxml>=5.0
python-multipart>=0.0.9

# Contact Enrichment
phonenumbers>=8.13.0

# SMS
twilio>=9.0.0

# OCR / image
pytesseract>=0.3.10
Pillow>=10.4.0
pdf2image>=1.17.0

# AI
pydantic-ai>=0.0.14
openai>=1.50.0
imagehash>=4.3.1
anthropic[vertex]>=0.40.0

# PDF / templating / auth
weasyprint>=61.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
shortuuid>=1.0.0
msal
pgvector
